refactor: changed structure of previous scraping keywords. Moved to separate file HTML.scala doc: added documentation for Scrape
