+++ src/main/scala/dsl/DSL.scala
@@ -1,57 +1,25 @@
 package org.unibo.scooby
 package dsl
 import core.scooby.Configuration
+import utility.document.html.HTMLElement
+import utility.document.{CommonHTMLExplorer, Document}
 
-import org.unibo.scooby.utility.document.html.HTMLElement
-
-import scala.compiletime.uninitialized
-import org.unibo.scooby.utility.document.{CommonHTMLExplorer, Document}
+import scala.annotation.targetName
 
 object DSL:
-
   export Config.*
   export Crawl.*
-  export Scrape.{scrape, matchesOf, select, elements, classes, attributes, and, id,
-                haveClass, haveId, haveTag, that, dont, including, or, rule}
+  export Scrape.scrape
   export Export.*
   export Utils.*
+  export HTML.*
 
   case class ScrapingResultSetting[T]()
 
   class ConfigurationBuilder[T](var configuration: Configuration[T], var scrapingResultSetting: ScrapingResultSetting[T]):
     def build: Configuration[T] = configuration
 
-  object Utils:
-
-    /**
-     * Retrieves tag from an [[HTMLElement]]
-     * @return the tag as [[String]].
-     */
-    def tag: HTMLElement => String = _.tag
-
-    /**
-     * Retrieves text from an [[HTMLElement]]
-     * @return text as [[String]].
-     */
-    def text: HTMLElement => String = _.text
-
-    /**
-     * Retrieves outer Html text from an [[HTMLElement]]
-     * @return outer Html text as [[String]].
-     */
-    def outerText: HTMLElement => String = _.outerHtml
-
-    /**
-     * Retrieves value of the specified attribute of an [[HTMLElement]].
-     * @return outer Html text as [[String]].
-     */
-
-    /**
-     * Retrieves value of the specified attribute of an [[HTMLElement]].
-     * @param attribute the attribute to read the value.
-     * @return the value of the attribute as [[String]].
-     */
-    def attr(attribute: String): HTMLElement => String = _.attr(attribute)
+  private object Utils:
 
     extension [T](x: Iterable[T])
 
@@ -62,6 +30,19 @@ object DSL:
        * @return the [[Iterable]] as result of the mapping function.
        */
       infix def get[A](f: T => A): Iterable[A] = x.map(f)
+
+    extension[T] (x: Iterable[T] )
+      infix inline def including[S >: T](y: Iterable[S]): Iterable[S] = x concat y
+
+      infix inline def that(predicate: T => Boolean): Iterable[T] = x filter predicate
+
+
+    extension[T] (x: T => Boolean)
+      infix def and(y: T => Boolean): T => Boolean = el => x(el) && y(el)
+      @targetName("and")
+      inline infix def &&(y: T => Boolean): T => Boolean = and(y)
+      infix def or(y: T => Boolean): T => Boolean = el => x(el) || y(el)
+      @targetName("and")
+      inline infix def ||(y: T => Boolean): T => Boolean = or(y)
     
 
-  def document[T <: Document & CommonHTMLExplorer](using document: T): T = document
